requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
pyotp
python-dotenv
NorenRestApiPy
websocket
requests
aiofiles
numba
mysql-connector-python